WIP DON'T MERGE Add a test for missing string documentation
THIS IS SUPPOSED TO FAIL AT THE MOMENT.

Documentation for all the strings is already written
in translatewiki (see #6457), but not yet imported into GitHub,
so it's supposed to fail now, which is good because you shouldn't trust
tests that never failed :)

Once all the documentation strings are imported, this pull request
should be rebased, and then it's supposed to pass.

The extra string in values-qq/string.xml is also there
just to check that it fails as it should, and will be removed.

Resolves #6485.